+++ aql/src/main/java/org/ehrbase/openehr/sdk/aql/dto/containment/ContainmentClassExpression.java
@@ -17,11 +17,23 @@
  */
 package org.ehrbase.openehr.sdk.aql.dto.containment;
 
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Objects;
+import org.apache.commons.collections4.CollectionUtils;
+import org.ehrbase.openehr.sdk.aql.dto.path.AndOperatorPredicate;
+import org.ehrbase.openehr.sdk.aql.serializer.PredicateDeserializer;
+import org.ehrbase.openehr.sdk.aql.serializer.PredicateSerializer;
 
-public class ContainmentClassExpression extends AbstractContainmentExpression {
+@JsonPropertyOrder({"type", "predicates", "contains"})
+public final class ContainmentClassExpression extends AbstractContainmentExpression {
 
     private String type;
+
+    protected List<AndOperatorPredicate> predicates;
 
     public String getType() {
         return type;
@@ -31,22 +43,42 @@ public class ContainmentClassExpression extends AbstractContainmentExpression {
         this.type = type;
     }
 
+    @JsonSerialize(using = PredicateSerializer.class)
+    @Override
+    public List<AndOperatorPredicate> getPredicates() {
+        return predicates;
+    }
+
+    @JsonDeserialize(using = PredicateDeserializer.class)
+    public void setPredicates(List<AndOperatorPredicate> predicates) {
+        if (predicates != null) {
+            this.predicates = new ArrayList<>(predicates);
+        } else {
+            this.predicates = null;
+        }
+    }
+
+    @Override
+    public boolean hasPredicates() {
+        return CollectionUtils.isNotEmpty(this.predicates);
+    }
+
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         if (!super.equals(o)) return false;
         ContainmentClassExpression that = (ContainmentClassExpression) o;
-        return Objects.equals(type, that.type);
+        return Objects.equals(type, that.type) && Objects.equals(predicates, that.predicates);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(super.hashCode(), type);
+        return Objects.hash(super.hashCode(), type, predicates);
     }
 
     @Override
     public String toString() {
-        return "ContainmentClassExpression{" + "type='" + type + '\'' + "} " + super.toString();
+        return "ContainmentClassExpression{type='%s', predicates=%s, %s}".formatted(type, predicates, super.toString());
     }
 }