+++ CHANGELOG.md
@@ -2,9 +2,24 @@
 
 Note: version releases in the 0.x.y range may introduce breaking changes.
 
+## [2.7.0]
+
+### Added
+
+- Added `QueryResponseData.meta` additional `fetch`, `offset` and `resultsize`
+  properties ([#559](https://github.com/ehrbase/openEHR_SDK/pull/559))
+
+### Changed
+
+- Sealed AqlQuery
+- AqlQuery: Support for VERSION without predicate
+
+### Fixed
+
 ## [2.6.0]
  ### Changed
 - Bumped libraries
+ ### Added 
  ### Fixed 
 - Parsing of string function CONTAINS ([5ef1eca](https://github.com/ehrbase/openEHR_SDK/commit/5ef1eca72289856fa94a04146aa876a4c5130423))
 - Add missing aggregate function name to the parser ([8fd08bb](https://github.com/ehrbase/openEHR_SDK/commit/8fd08bbd1ce3b5a217cc8e2b701e32c2125bf0bf))
@@ -360,3 +375,4 @@ Note: version releases in the 0.x.y range may introduce breaking changes.
 [2.4.0]: https://github.com/ehrbase/openEHR_SDK/compare/v2.3.0...v2.4.0
 [2.5.0]: https://github.com/ehrbase/openEHR_SDK/compare/v2.4.0...v2.5.0
 [2.6.0]: https://github.com/ehrbase/openEHR_SDK/compare/v2.5.0...v2.6.0
+[2.7.0]: https://github.com/ehrbase/openEHR_SDK/compare/v2.6.0...v2.7.0